refactor(api-client): add runs namespace
