+++ docs/api/api-client.md
@@ -89,12 +89,12 @@ for (const [step, state] of Object.entries(setupProgress.steps)) {
 
 ## Read project run history
 
-Use `project.getRunHistory()` to inspect historical Data Mart executions visible to the current
+Use `runs.getHistory()` to inspect historical Data Mart executions visible to the current
 project member. Pass optional `limit` and `offset` values to page through the project-wide history;
 the API defaults to at most 100 runs.
 
 ```ts
-const history = await client.project.getRunHistory({ limit: 50, offset: 0 });
+const history = await client.runs.getHistory({ limit: 50, offset: 0 });
 
 for (const run of history.runs) {
   console.log(run.dataMart.title, run.type, run.status, run.finishedAt);
